requirements.txt
amqp
asgiref
async-timeout
attrs
beautifulsoup4
billiard
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cron-descriptor
Django
django-celery-beat
django-timezone-field
djangorestframework
djangorestframework-simplejwt
exceptiongroup
fake-useragent
fuzzywuzzy
gunicorn
h11
idna
importlib-resources
kombu
lxml
mysqlclient
outcome
packaging
prompt-toolkit
PyJWT
PySocks
python-crontab
python-dateutil
pytz
redis
requests
selenium
six
sniffio
sortedcontainers
soupsieve
sqlparse
trio
trio-websocket
typing-extensions
tzdata
Unidecode
urllib3
vine
wcwidth
websocket-client
wsproto
zipp
eventlet
python-Levenshtein
